feat: backend llm client 코드 정리
