+++ backend/requirements.txt
@@ -18,7 +18,6 @@ pip==25.1.1
 platformdirs==4.3.8
 pluggy==1.6.0
 pycodestyle==2.14.0
-pydantic==2.11.7
 pydantic-core==2.33.2
 pyflakes==3.4.0
 pygments==2.19.2
@@ -29,3 +28,10 @@ starlette==0.46.2
 typing-extensions==4.14.0
 typing-inspection==0.4.1
 uvicorn==0.35.0
+pydantic==2.7.1
+langchain==0.1.16
+langchain-core==0.1.53
+langchain-community==0.0.38
+langchain-google-vertexai==1.0.2
+langchain-text-splitters==0.0.2
+google-cloud-aiplatform==1.51.0
